i need help with neuro evolution

+++ flappybird/Constants.java
@@ -17,8 +17,8 @@ public class Constants {
     public static final int EYEOFFSET_Y = 4;
     public static final int EYE_RADIUS = 2;
 
-    public static final int GRAVITY = 1000; // acceleration constant (UNITS: pixels/s^2)
-    public static final int REBOUND_VELOCITY = -350; // initial jump velocity (UNITS: pixels/s)
+    public static final double GRAVITY = 1000; // acceleration constant (UNITS: pixels/s^2)
+    public static final double REBOUND_VELOCITY = -350; // initial jump velocity (UNITS: pixels/s)
     public static final double DURATION = 0.016; // KeyFrame duration (UNITS: s)
 
     public static final int PIPE_WIDTH = 50;
